Moving channel_dim into channel_props.
