+++ toqito/channel_ops/choi_to_kraus.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from toqito.helper import channel_dim
+from toqito.channel_props import channel_dim
 from toqito.matrix_ops import unvec
 from toqito.matrix_props import is_hermitian, is_positive_semidefinite
