+++ toqito/channel_ops/dual_channel.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from toqito.helper import channel_dim
+from toqito.channel_props import channel_dim
 from toqito.perms import swap
