+++ toqito/channel_props/__init__.py
@@ -1,5 +1,6 @@
 """A number of properties of quantum channels."""
 
+from toqito.channel_props.channel_dim import channel_dim
 from toqito.channel_props.is_herm_preserving import is_herm_preserving
 from toqito.channel_props.is_completely_positive import is_completely_positive
 from toqito.channel_props.is_positive import is_positive
