+++ toqito/channel_props/is_unital.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 from toqito.channel_ops import apply_channel
-from toqito.helper import channel_dim
+from toqito.channel_props import channel_dim
 from toqito.matrix_props import is_identity
