+++ toqito/channel_props/tests/test_channel_dim.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pytest
 
-from toqito.helper import channel_dim
+from toqito.channel_props import channel_dim
 from toqito.matrices import pauli
 from toqito.perms import swap_operator
