+++ toqito/helper/__init__.py
@@ -5,4 +5,3 @@ from toqito.helper.np_array_as_expr import np_array_as_expr
 from toqito.helper.update_odometer import update_odometer
 from toqito.helper.cvx_kron import cvx_kron
 from toqito.helper.npa_hierarchy import npa_constraints
-from toqito.helper.channel_dim import channel_dim
